refactor: 초성, 자소 변환 로직 추상화 의존하도록 리팩토링
