+++ src/main/java/com/joojoo/api/common/codeRandom/GeneratorRandom.java
@@ -1,4 +1,4 @@
-package com.joojoo.api.util.port.in;
+package com.joojoo.api.common.codeRandom;
 
 public interface GeneratorRandom {
     String getRandomName();
