+++ src/main/java/com/joojoo/api/common/hangul/HangulConverter.java
@@ -1,0 +1,6 @@
+package com.joojoo.api.common.hangul;
+
+public interface HangulConverter {
+    String jasoConvert(String text);
+    String chosungConvert(String text);
+}
