+++ src/main/java/com/joojoo/api/common/search/SearchRangeFactory.java
@@ -1,7 +1,7 @@
 package com.joojoo.api.common.search;
 
 import com.joojoo.api.common.search.range.SearchRange;
-import com.joojoo.api.util.port.in.JasoConverter;
+import com.joojoo.api.common.hangul.HangulConverter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -9,16 +9,16 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class SearchRangeFactory {
 
-    private final JasoConverter jasoConverter;
+    private final HangulConverter hangulConverter;
     private static final String END_UNICODE = "\uffff";
 
     public SearchRange createSearchRange(String query) {
-        String converted = jasoConverter.convert(query);
+        String converted = hangulConverter.jasoConvert(query);
         return buildRange(converted);
     }
 
     public SearchRange createSearchTagKeyRange(String query, Long userId) {
-        String converted = jasoConverter.convert(query);
+        String converted = hangulConverter.jasoConvert(query);
         String convertedQuery = userId + ":" + converted;
         return buildRange(convertedQuery);
     }
