--- src/main/java/com/joojoo/api/util/port/in/JasoConverter.java
@@ -1,5 +1,0 @@
-package com.joojoo.api.util.port.in;
-
-public interface JasoConverter {
-    String convert(String text);
-}
